Rework handling of property changes to connman service objects
It's really needed in a lot cases to just parse the properties and store the result of a
changed property in the service object and then forward processing to connected handlers.
This reworks the property change notification handling so a possible user can register can
register for property changes and will be notified whenever any property of the service
object changes.

In the case of the already existing state notification handler the new structure is used
too but the client is restricted to just use changes to the State property of the service
object.

Open-webOS-DCO-1.0-Signed-off-by: Simon Busch <morphis@gravedo.de>